requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
grpcio==1.14.2
grpcio-tools==1.14.2
pkg-resources==0.0.0
protobuf==3.9.1
six==1.12.0
Pillow==6.1.0
azure-storage==0.36.0